+++ magentoerpconnect/CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 * New helper in importer to import dependencies
 * allow to customize the available versions without overriding the 'version' field
 * Products deactivated on Magento are imported as not sellable / purchasable
+* New option 'Create Invoice On' on payment methods with options 'on paid', 'on validate'
 
 
 2.3.1 (2014-01-23)